more unit tests, input handler improvement, testing on hff data
