+++ src/tests/mod.rs
@@ -1,1 +1,3 @@
 pub mod backtester_test;
+pub mod input_handler_test;
+pub mod metrics_test;
